change ember-cli-photoswipe to ember-photoswipe.
ember-cli-photoswipe addon causing a bug that when click on a thumbnail to goto a sub-album, the action is to open photoswipe, which i don't want to.
ember-photoswipe gives a change to define my own action to every thumbnail, i can choose to open sub-album or open photoswipe, or, display comments.

* removed ember-welcome-page, ember-cli-photoswipe
* removed bower.json
* added ember-cli-sass